requirements.txt
faster-whisper==1.0.3
ffmpeg-python==0.2.0
pyannote.audio>=3.1,<4.0
